refactor: move statusbar to workspace/statusbar
