+++ packages/base/src/workspace/statusbar/SpectaPresentationProgressBar.tsx
