+++ packages/base/src/workspace/statusbar/StatusBar.tsx
